+++ .env
@@ -7,6 +7,6 @@ VUE_APP_I18N_FALLBACK_LOCALE=en
 APP_GEMINI_API_KEY=your_gemini_key
 
 # model name to use for translation https://cloud.google.com/vertex-ai/generative-ai/docs/learn/model-versions#latest-stable
-FIX_LOCALES_MODEL_NAME=gemini-2.0-flash
+FIX_LOCALES_MODEL_NAME=gemini-2.5-flash-lite
 # model temperature - 0.0 means no randomness
 FIX_LOCALES_TEMPERATURE=0.0
