+++ client-app/pages/account/promotion-coupons.vue
@@ -5,50 +5,16 @@
     </VcTypography>
 
     <div v-if="result?.promotionCoupons?.items" class="promotion-coupons__wrapper">
-      <div v-for="coupon in result.promotionCoupons.items" :key="coupon.id" class="coupon-item">
-        <div class="coupon-item__title">
-          <p>{{ coupon.systemName }}</p>
-        </div>
-
-        <div class="coupon-item__name">
-          <p>{{ coupon.name }}</p>
-        </div>
-
-        <div class="coupon-item__description">
-          <p>{{ coupon.description }}</p>
-        </div>
-
-        <div class="coupon-item__end-date">
-          <template v-if="coupon.endDate">
-            <span class="coupon-item__end-date-label">Expired</span>
-
-            <span class="coupon-item__end-date-value">{{ $d(coupon.endDate, "short") }}</span>
-          </template>
-        </div>
-
-        <button type="button" class="coupon-item__code" @click="copyCoupon(coupon.couponCode)">
-          <span class="coupon-item__code-value">{{ coupon.couponCode }}</span>
-
-          <span class="coupon-item__code-label">Click to copy</span>
-        </button>
-      </div>
+      <CouponItem v-for="coupon in result.promotionCoupons.items" :key="coupon.id" :coupon="coupon" />
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
-import { useClipboard } from "@vueuse/core";
 import { getPromotionCoupons } from "@/core/api/graphql";
+import { CouponItem } from "@/shared/account";
 
 const { result } = getPromotionCoupons();
-
-const { copy, isSupported } = useClipboard();
-
-async function copyCoupon(text?: string) {
-  if (isSupported && text?.trim()) {
-    await copy(text);
-  }
-}
 </script>
 
 <style lang="scss">
