+++ client-app/shared/account/components/index.ts
@@ -1,6 +1,7 @@
 export { default as AccountShell } from "./account-shell.vue";
 export { default as AddressDropdownMenu } from "./address-dropdown-menu.vue";
 export { default as ChangePasswordForm } from "./change-password-form.vue";
+export { default as CouponItem } from "./coupon-item.vue";
 export { default as CreditCard } from "./credit-card.vue";
 export { default as CreditCardSkeleton } from "./credit-card-skeleton.vue";
 export { default as ForgotPasswordForm } from "./forgot-password-form.vue";
