+++ locales/en.json
@@ -543,6 +543,11 @@
         "end_date_label": "End date",
         "apply_button": "@:common.buttons.apply",
         "reset_button": "@:common.buttons.reset"
+      },
+      "promotion_coupons": {
+        "expired": "Expired",
+        "click_to_copy": "Click to copy",
+        "clipboard_success": "Coupon copied successfully"
       }
     },
     "company": {
