+++ locales/no.json
@@ -543,6 +543,11 @@
         "end_date_label": "Sluttdato",
         "apply_button": "@:common.buttons.apply",
         "reset_button": "@:common.buttons.reset"
+      },
+      "promotion_coupons": {
+        "expired": "Utløpt",
+        "click_to_copy": "Klikk for å kopiere",
+        "clipboard_success": "Kupong kopiert vellykket"
       }
     },
     "company": {
@@ -1268,6 +1273,9 @@
         "title": "Identitetsetterligning",
         "success": "Identitetsetterligning vellykket",
         "error": "Problemer med identitetsetterligning. Kontroller tilgangsrettighetene dine"
+      },
+      "promotion_coupons": {
+        "title": "Alle kuponger og kampanjer"
       }
     },
     "company": {
