+++ locales/pt.json
@@ -543,6 +543,11 @@
         "end_date_label": "Data de término",
         "apply_button": "@:common.buttons.apply",
         "reset_button": "@:common.buttons.reset"
+      },
+      "promotion_coupons": {
+        "expired": "Expirado",
+        "click_to_copy": "Clique para copiar",
+        "clipboard_success": "Cupom copiado com sucesso"
       }
     },
     "company": {
@@ -1268,6 +1273,9 @@
         "title": "Impersonação",
         "success": "Impersonificação bem-sucedida",
         "error": "Alguns problemas durante a personificação. Verifique suas permissões"
+      },
+      "promotion_coupons": {
+        "title": "Todos os cupons e promoções"
       }
     },
     "company": {
