+++ locales/zh.json
@@ -543,6 +543,11 @@
         "end_date_label": "结束日期",
         "apply_button": "@:common.buttons.apply",
         "reset_button": "@:common.buttons.reset"
+      },
+      "promotion_coupons": {
+        "expired": "已过期",
+        "click_to_copy": "点击复制",
+        "clipboard_success": "优惠券已复制"
       }
     },
     "company": {
@@ -1268,6 +1273,9 @@
         "title": "冒充",
         "success": "冒充成功",
         "error": "冒充时出现一些问题。检查您的权限"
+      },
+      "promotion_coupons": {
+        "title": "所有优惠券与促销"
       }
     },
     "company": {
